refactor: split monolithic app.rs into focused modules (#18)
* refactor: remove monolithic app.rs in preparation for modularization

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)


* refactor: split app.rs into focused modules

Split 3102-line monolithic app.rs into fine-grained modules:

- state.rs: dialog states, enums, view modes

- messages.rs: TEA Message enum

- runtime.rs: RecoveryRuntime and CreateTaskRuntime traits

- polling.rs: async status poller thread

- dialogs.rs: dialog key event handlers

- actions.rs: task/category operations

- mod.rs: core App struct and main logic

All modules now under 2000 lines (largest: mod.rs at 1563 lines)

Ultraworked with Sisyphus


---------
