+++ src/app/messages.rs
@@ -1,0 +1,65 @@
+//! TEA (The Elm Architecture) messages for application events
+
+use std::path::PathBuf;
+
+use crossterm::event::{KeyEvent, MouseEvent};
+
+use super::state::{CategoryInputField, DeleteTaskField, NewProjectField, NewTaskField};
+
+#[derive(Debug, Clone, Eq, PartialEq)]
+pub enum Message {
+    Key(KeyEvent),
+    Mouse(MouseEvent),
+    Tick,
+    Resize(u16, u16),
+    NavigateLeft,
+    NavigateRight,
+    SelectUp,
+    SelectDown,
+    AttachSelectedTask,
+    OpenNewTaskDialog,
+    OpenCommandPalette,
+    OpenProjectList,
+    DismissDialog,
+    FocusColumn(usize),
+    SelectTask(usize, usize),
+    SelectTaskInSidePanel(usize),
+    OpenAddCategoryDialog,
+    OpenRenameCategoryDialog,
+    OpenDeleteCategoryDialog,
+    OpenDeleteTaskDialog,
+    SubmitCategoryInput,
+    ConfirmDeleteCategory,
+    MoveTaskLeft,
+    MoveTaskRight,
+    MoveTaskUp,
+    MoveTaskDown,
+    CreateTask,
+    DeleteTaskToggleKillTmux,
+    DeleteTaskToggleRemoveWorktree,
+    DeleteTaskToggleDeleteBranch,
+    ConfirmDeleteTask,
+    WorktreeNotFoundRecreate,
+    WorktreeNotFoundMarkBroken,
+    RepoUnavailableDismiss,
+    ConfirmQuit,
+    CancelQuit,
+    ExecuteCommand(String),
+    CycleCategoryColor(usize),
+    SwitchToProjectList,
+    SwitchToBoard(PathBuf),
+    ProjectListSelectUp,
+    ProjectListSelectDown,
+    ProjectListConfirm,
+    OpenNewProjectDialog,
+    CreateProject,
+    FocusNewTaskField(NewTaskField),
+    ToggleNewTaskCheckbox,
+    FocusCategoryInputField(CategoryInputField),
+    FocusNewProjectField(NewProjectField),
+    FocusDeleteTaskField(DeleteTaskField),
+    ToggleDeleteTaskCheckbox(DeleteTaskField),
+    FocusDialogButton(String),
+    SelectProject(usize),
+    SelectCommandPaletteItem(usize),
+}
